중복 코드들 정리
